requirements.txt
# Core
langchain==1.1.0
langchain-core==1.1.0

# OpenAI Integration
langchain-openai==1.1.0
openai>=1.0.0

# Anthropic Integration
langchain-anthropic==1.1.0

# Google Gemini (PaLM) Integration
# pin to a known recent release to avoid pip pulling the 0.0.1 broken package
langchain-google-genai==3.2.0
google-genai==1.52.0

# Hugging Face Integration
langchain-huggingface==1.1.0
transformers>=4.30.0
huggingface-hub>=0.14.0

# Environment / Utils
python-dotenv>=1.0.0

# Embeddings (REQUIRED for all-MiniLM-L6-v2)
sentence-transformers>=2.6.0

# ML / numeric
numpy>=1.24.0
scikit-learn>=1.2.0
torch>=2.0.0
